+++ src/haverscript/__init__.py
@@ -1,3 +1,23 @@
-from .haverscript import Echo, Model, Response, accept, connect, fresh, valid_json
+from .haverscript import (
+    Configuration,
+    Echo,
+    Host,
+    Model,
+    Response,
+    accept,
+    connect,
+    fresh,
+    valid_json,
+)
 
-__all__ = ["Echo", "Model", "Response", "accept", "connect", "fresh", "valid_json"]
+__all__ = [
+    "Configuration",
+    "Echo",
+    "Host",
+    "Model",
+    "Response",
+    "accept",
+    "connect",
+    "fresh",
+    "valid_json",
+]